+++ packages/dev-server/index.js
@@ -14,7 +14,7 @@ devServer.reactHot = reactHot
  * @param {bool}   [options.historyApiFallback]
  * @param {bool}   [options.hot]
  * @param {bool}   [options.inline]
- * @param {string|string[]} [entry]   Defaults to 'webpack/hot/only-dev-server'
+ * @param {string|string[]} [entry]
  * @return {Function}
  */
 function devServer (options, entry) {
@@ -40,14 +40,13 @@ function devServer (options, entry) {
 }
 
 function postConfig (context, util) {
-  const entryPointsToAdd = context.devServer.entry.length > 0
-    ? context.devServer.entry
-    : [ 'webpack/hot/only-dev-server' ]
+  const entryPointsToAdd = context.devServer.entry
 
   return prevConfig => {
     return util.merge({
       devServer: Object.assign({
         hot: true,
+        hotOnly: true,
         historyApiFallback: true,
         inline: true
       }, context.devServer.options),