+++ packages/webpack/__e2e-fixtures__/elm/main.js
@@ -1,0 +1,3 @@
+const Elm = require('./app')
+
+Elm.App.embed(document.body)