+++ packages/webpack/__tests__/end-to-end.test.js
@@ -76,6 +76,18 @@ test('the postcss/sass/source-maps project build does not fail', async (t) => {
   await runWebpack(config)
 })
 
+test('building the elm project works', async (t) => {
+  const projectPath = path.join(fixturesPath, 'elm')
+  const buildPath = path.join(projectPath, 'build')
+
+  const config = require(path.join(projectPath, 'webpack.config.js'))
+  await runWebpack(config)
+
+  global.window = await setUpJsdomEnv()
+  global.document = global.window.document
+  require(path.join(buildPath, 'bundle.js'))
+})
+
 function runWebpack (config) {
   return new Promise((resolve, reject) => {
     webpack(config, (error, stats) => {