+++ packages/webpack/__tests__/integration.test.js
@@ -80,12 +80,11 @@ test('complete webpack config creation', (t) => {
     } ]
   })
 
-  t.deepEqual(webpackConfig.entry, {
-    main: [ './src/main.js', 'webpack/hot/only-dev-server' ]
-  })
+  t.deepEqual(webpackConfig.entry, { main: [ './src/main.js' ] })
 
   t.deepEqual(webpackConfig.devServer, {
     hot: true,
+    hotOnly: true,
     historyApiFallback: true,
     inline: true,
     proxy: {