+++ packages/typescript/CHANGELOG.md
@@ -1,8 +1,12 @@
 # @webpack-blocks/typescript - Changelog
 
-## 2.0.0
+## 1.0
 
 - Updated for new core API ([#125](https://github.com/andywer/webpack-blocks/issues/125))
+
+## 0.4.1
+
+- Allow passing custom loader options (see [#141](https://github.com/andywer/webpack-blocks/pull/141))
 
 ## 0.4.0